pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_backend"

[project]
name = "steering-reliability"
version = "0.1.0"
description = "Research pipeline for measuring steering direction reliability under distribution shift"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "MATS Research"}
]
dependencies = [
    "torch>=2.0.0",
    "transformer-lens>=1.14.0",
    "transformers>=4.30.0",
    "datasets>=2.14.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "pyarrow>=12.0.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.0.280",
    "ipython>=8.14.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.24.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"
